pyproject.toml
[tool.poetry]
name = "teste-agriness"
version = "0.1.0"
description = ""
authors = ["sergiomachadopk <sergio.machadopk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.1"
psycopg2-binary = "^2.9.6"
pytest = "^7.3.1"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.2"
dj-database-url = "^2.0.0"
python-decouple = "^3.8"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
